requirements.txt
asgiref==3.3.1
certifi==2020.12.5
Django==3.1.5
django-filter==2.4.0
djangorestframework==3.12.2
Markdown==3.3.3
psycopg2-binary==2.8.6
Pygments==2.7.4
python-dotenv==0.15.0
pytz==2020.5
sqlparse==0.4.1
